+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -185,10 +185,10 @@ import javax.tools.Diagnostic.Kind;
     // See Issue 979.
     "conservativeUninferredTypeArguments",
 
-    // Issues a "redundant.anno" warning if the annotation written on the type is the same as
-    // the default annotation would be applied on the type. When not supplying this flag, this
+    // Issues a "redundant.anno" warning if the annotation explicitly written on the type is
+    // the same as the default annotation for this type and location. Without this flag, this
     // option is disabled.
-    "checkRedundantAnnotations",
+    "warnRedundantAnnotations",
 
     // Whether to ignore all subtype tests for type arguments that
     // were inferred for a raw type. Defaults to true.